refactor: DispatcherController 리팩토링
기존에 DispatcherController에게 불필요하게 위임되었던 책임을 각 도메인 별 컨트롤러에게 나눴습니다.
