+++ src/main/java/kr/co/programmers/springbootbasic/customer/service/BlackCustomerService.java
@@ -3,7 +3,6 @@ package kr.co.programmers.springbootbasic.customer.service;
 import kr.co.programmers.springbootbasic.customer.domain.Customer;
 import kr.co.programmers.springbootbasic.customer.dto.CustomerResponse;
 import kr.co.programmers.springbootbasic.customer.repository.BlackCustomerRepository;
-import kr.co.programmers.springbootbasic.util.ApplicationUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Profile;
@@ -21,7 +20,7 @@ public class BlackCustomerService {
         this.blackCustomerRepository = userRepository;
     }
 
-    public List<CustomerResponse> listAllBlackCustomer() {
+    public List<CustomerResponse> findAllBlackCustomer() {
         logger.info("블랙리스트 고객들을 조회합니다...");
         List<Customer> userList = blackCustomerRepository.listAllBlackCustomer();
